feat: Add management of source control
Both applications and organizations can manage the source control
independently or by inheriting

In the tests I have added this repository as the repository_url as
you have to use a URI that matches a provider. In the IQ instance
I was using the provider at the root organisation is set to GitHub,
this might mean the acceptance tests I have added here may fail for
others with a different configuration.

Signed-off-by: grahamhar <grahamhar@users.noreply.github.com>